Testing the storing of the Path parts and the Form data. Some variables can be moved from Request to RequestData. Autoindex content generator needs to be called, when a folder is found

+++ working_kqueue_cpp/Server.hpp
@@ -23,7 +23,7 @@ class Server
 	private:
 		int _kq;
 		struct addrinfo *_addr;
-		int _listening_socket;		
+		size_t _listening_socket;		
 
 	public:
 		Server();
